modify defines & variables for rotation/hooks

+++ include/defines.h
@@ -3,8 +3,8 @@
 
 # define WIDTH 1280
 # define HEIGHT 720
-# define ROTATION_SPEED 0.0174
-# define MOVE_SPEED 0.01
+# define ROTATION_SPEED (double)0.0174
+# define MOVE_SPEED (double)0.01
 
 # define BPP sizeof(int32_t)
